fix(keysign): guard embedded QR logo scaling against null color space
GenerateQrBitmap scaled the embedded logo with raw `logo.scale()`, which
throws "can't create bitmap without a color space" on devices that hand
back a null-color-space bitmap from decodeResource. ShareVaultQrViewModel
catches only WriterException, so the #5089 crash survived on this path.

Extract the color-space-safe scale helper into a shared file and reuse it
for the logo in GenerateQrBitmap.
